Move QueryHorizon to IR module
Also move all implementing classes
