+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/ProcedureCallProjection.scala
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2002-2017 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.cypher.internal.compiler.v3_2.planner
+
+import org.neo4j.cypher.internal.compiler.v3_2.ast.ResolvedCall
+import org.neo4j.cypher.internal.compiler.v3_2.spi.ProcedureReadOnlyAccess
+import org.neo4j.cypher.internal.ir.v3_2._
+
+case class ProcedureCallProjection(call: ResolvedCall) extends QueryHorizon {
+  override def exposedSymbols(coveredIds: Set[IdName]): Set[IdName] = coveredIds ++ call.callResults.map { result => IdName.fromVariable(result.variable) }
+
+  override def dependingExpressions = call.callArguments
+
+  override def preferredStrictness = call.signature.accessMode match {
+    case _:ProcedureReadOnlyAccess => Some(LazyMode)
+    case _ => Some(EagerMode)
+  }
+}
