+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/OptionalMatchRemover.scala
@@ -19,15 +19,15 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_2.planner.logical
 
-import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase
-import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase.LOGICAL_PLANNING
 import org.neo4j.cypher.internal.compiler.v3_2.phases.{CompilationState, CompilerContext}
-import org.neo4j.cypher.internal.compiler.v3_2.planner.{AggregatingQueryProjection, QueryGraph, RegularPlannerQuery, _}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.{QueryGraph, RegularPlannerQuery, _}
 import org.neo4j.cypher.internal.frontend.v3_2.Rewritable._
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{Expression, FunctionInvocation, _}
-import org.neo4j.cypher.internal.frontend.v3_2.phases.{Condition, Phase}
+import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase
+import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase.LOGICAL_PLANNING
+import org.neo4j.cypher.internal.frontend.v3_2.phases.Phase
 import org.neo4j.cypher.internal.frontend.v3_2.{InputPosition, Rewriter, topDown}
-import org.neo4j.cypher.internal.ir.v3_2.{IdName, PatternRelationship}
+import org.neo4j.cypher.internal.ir.v3_2.{AggregatingQueryProjection, IdName, PatternRelationship}
 
 import scala.annotation.tailrec
 import scala.collection.mutable.ListBuffer
@@ -37,7 +37,7 @@ case object OptionalMatchRemover extends PlannerQueryRewriter {
 
   override def description: String = "remove optional match when possible"
 
-  override def postConditions: Set[Condition] = Set.empty
+  override def postConditions = Set.empty
 
   override def instance(ignored: CompilerContext): Rewriter = topDown(Rewriter.lift {
     case RegularPlannerQuery(graph, proj@AggregatingQueryProjection(distinctExpressions, aggregations, _), tail)
