+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/PlanEventHorizon.scala
@@ -22,6 +22,7 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner.logical
 import org.neo4j.cypher.internal.compiler.v3_2.planner._
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans.LogicalPlan
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.steps.{aggregation, projection, sortSkipAndLimit}
+import org.neo4j.cypher.internal.ir.v3_2._
 
 /*
 Planning event horizons means planning the WITH clauses between query patterns. Some of these clauses are inlined
