+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/steps/countStorePlanner.scala
@@ -24,14 +24,14 @@ import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.LogicalPlanningCo
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans._
 import org.neo4j.cypher.internal.frontend.v3_2.SemanticDirection.{INCOMING, OUTGOING}
 import org.neo4j.cypher.internal.frontend.v3_2.ast._
-import org.neo4j.cypher.internal.ir.v3_2.{IdName, PatternRelationship, SimplePatternLength}
+import org.neo4j.cypher.internal.ir.v3_2.{AggregatingQueryProjection, IdName, PatternRelationship, SimplePatternLength}
 
 case object countStorePlanner {
 
   def apply(query: PlannerQuery)(implicit context: LogicalPlanningContext): Option[LogicalPlan] = {
     implicit val semanticTable = context.semanticTable
     query.horizon match {
-      case AggregatingQueryProjection(groupingKeys, aggregatingExpressions, shuffle)
+      case AggregatingQueryProjection(groupingKeys, aggregatingExpressions, _)
         if groupingKeys.isEmpty && aggregatingExpressions.size == 1 =>
         val (columnName, exp) = aggregatingExpressions.head
         val countStorePlan = checkForValidQueryGraph(query, columnName, exp)
