+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/steps/sortSkipAndLimit.scala
@@ -21,10 +21,10 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.steps
 
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical._
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans._
-import org.neo4j.cypher.internal.compiler.v3_2.planner.{PlannerQuery, QueryProjection}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.PlannerQuery
 import org.neo4j.cypher.internal.frontend.v3_2.ast.Variable
 import org.neo4j.cypher.internal.frontend.v3_2.{InternalException, ast}
-import org.neo4j.cypher.internal.ir.v3_2.IdName
+import org.neo4j.cypher.internal.ir.v3_2.{IdName, QueryProjection}
 
 object sortSkipAndLimit extends PlanTransformer[PlannerQuery] {
