+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/steps/SortSkipAndLimitTest.scala
@@ -25,7 +25,7 @@ import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.{Ascending, Logic
 import org.neo4j.cypher.internal.frontend.v3_2.ast
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{AscSortItem, PatternExpression}
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
-import org.neo4j.cypher.internal.ir.v3_2.{Cardinality, IdName, QueryShuffle}
+import org.neo4j.cypher.internal.ir.v3_2.{Cardinality, IdName, QueryShuffle, RegularQueryProjection}
 
 class SortSkipAndLimitTest extends CypherFunSuite with LogicalPlanningTestSupport {
